Show context bottom sheet + add reply text field to conversation.

+++ feature/metis/src/main/kotlin/de/tum/informatics/www1/artemis/native_app/feature/metis/model/MetisContext.kt
@@ -40,6 +40,6 @@ sealed class MetisContext(
 
     @Serializable
     @SerialName("conversation")
-    data class Conversation(override val courseId: Long, val conversationId: Int) :
+    data class Conversation(override val courseId: Long, val conversationId: Long) :
         MetisContext(listOf(), "messages", "answer-messages"), ConversationMetisContext
 }
